vibe the code into a better structure
