+++ examples/example.lask-config
@@ -3,7 +3,7 @@
 
 [default]
 # The provider to use for LLM requests
-# Options: openai, anthropic, aws
+# Options: openai, anthropic, aws, azure
 provider = openai
 
 # Default API key (you can also set provider-specific keys in each section)
@@ -42,6 +42,20 @@ model = claude-3-opus-20240229
 
 # temperature = 0.7
 # max_tokens = 4096
+
+# Azure OpenAI configuration
+[azure]
+# Your Azure OpenAI API key. If not specified, falls back to the default api_key
+# api_key = your-azure-openai-api-key-here
+
+# Required Azure OpenAI settings
+resource_name = your-resource-name
+deployment_id = your-deployment-id
+
+# Optional Azure OpenAI settings
+# api_version = 2023-05-15
+# temperature = 0.7
+# max_tokens = 2000
 
 # AWS Bedrock configuration
 [aws]
