+++ setup.cfg
@@ -9,10 +9,15 @@ classifiers =
     Operating System :: OS Independent
 
 [options]
-py_modules = lask
+packages = find:
+package_dir =
+    = .
 install_requires =
     requests
 
 [options.entry_points]
 console_scripts =
     lask = lask:main
+
+[options.packages.find]
+include = src*
