+++ src/__init__.py
@@ -1,1 +1,9 @@
-# Empty file to make this directory a Python package
+"""
+Package initialization for lask.
+
+This module provides access to the main lask functionality.
+"""
+
+from src.main import main
+
+__all__ = ["main"]
